Update test naming
